Improved fps controls, octree graph. Has issues!
There are still a number of issues, but this basically demo-able.
The octree sped up raycasts *significantly*.
Still has noticeable frame rate drops in certain areas... might be when
the octree is switching between nodes?
You can get stuck in parts of the geometry because I think my algorithm
is incorrect (ie when checking for collisions in x and z directions)!
Needs tweaking for movement speed and gravity, player height should be
configurable from some other module (probably), it's still possible to
throw yourself into the abyss. I tried preventing that, but it had some
issues.