More collision fixes
The new collision is more simple AND robust! I've tested with many of
the walls in our geometry; it *mostly* seems impossible to phase through
walls with at least one (probably more, but I didn't find any) annoying
exception. The back wall of the starting room can be walked right
through most of the time. More strangely, it's a one-way trip. Might be
an octree quirk. On the bright side, it's not a situation in which one
can become permanently *stuck*, so that's nice.